put arrays in the code, but we won't use them. just for cool
To actually implement portals, you only have to remove the element and ADD it at the index it enters.
So if you add it to index i from the right, you first remove the digit d from the number,
and then to the new number add (d * 10^i).

It remains to be seen how this works if the number's most significant digit extends beyond the entering
portal (probably just breaks), but for now, we can guarantee that the number's length will be, at rest,
lower than needed to enter the portal.